Split sources into language-agnostic and Java-specific
