+++ model/src/main/java/io/spine/validation/DefaultValue.java
@@ -24,27 +24,18 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.kanban.codegen;
+package io.spine.validation;
 
+import com.google.protobuf.ByteString;
 import io.spine.protodata.Field;
 import io.spine.protodata.PrimitiveType;
 import io.spine.protodata.Type;
+import io.spine.protodata.Type.KindCase;
 import io.spine.protodata.TypeName;
-import io.spine.validation.EnumValue;
-import io.spine.validation.ListValue;
-import io.spine.validation.MapValue;
-import io.spine.validation.MessageValue;
-import io.spine.validation.Value;
+import io.spine.util.Exceptions;
 import org.jetbrains.annotations.NotNull;
 
 import static com.google.common.base.Preconditions.checkNotNull;
-import static com.google.protobuf.ByteString.EMPTY;
-import static io.spine.protodata.PrimitiveType.PT_UNKNOWN;
-import static io.spine.protodata.PrimitiveType.TYPE_BYTES;
-import static io.spine.protodata.PrimitiveType.TYPE_STRING;
-import static io.spine.protodata.PrimitiveType.UNRECOGNIZED;
-import static io.spine.util.Exceptions.newIllegalArgumentException;
-import static io.spine.util.Exceptions.newIllegalStateException;
 
 /**
  * A factory of default values of Protobuf message fields.
@@ -79,7 +70,7 @@ public final class DefaultValue {
     }
 
     private static Value defaultValue(Type type) {
-        Type.KindCase kind = type.getKindCase();
+        KindCase kind = type.getKindCase();
         switch (kind) {
             case MESSAGE:
                 return messageValue(type);
@@ -96,20 +87,20 @@ public final class DefaultValue {
     @NotNull
     private static Value primitiveValue(Type type) {
         PrimitiveType primitiveType = type.getPrimitive();
-        if (primitiveType == PT_UNKNOWN || primitiveType == UNRECOGNIZED) {
-            throw newIllegalArgumentException("Unknown type `%s`.", primitiveType);
+        if (primitiveType == PrimitiveType.PT_UNKNOWN || primitiveType == PrimitiveType.UNRECOGNIZED) {
+            throw Exceptions.newIllegalArgumentException("Unknown type `%s`.", primitiveType);
         }
-        if (primitiveType == TYPE_STRING) {
+        if (primitiveType == PrimitiveType.TYPE_STRING) {
             return Value.newBuilder()
                         .setStringValue("")
                         .vBuild();
         }
-        if (primitiveType == TYPE_BYTES) {
+        if (primitiveType == PrimitiveType.TYPE_BYTES) {
             return Value.newBuilder()
-                        .setBytesValue(EMPTY)
+                        .setBytesValue(ByteString.EMPTY)
                         .vBuild();
         }
-        throw newIllegalStateException(
+        throw Exceptions.newIllegalStateException(
                 "Fields of type `%s` do not support `(required)` validation.",
                 primitiveType
         );
