+++ model/src/main/java/io/spine/validation/MessageValidationRepository.java
@@ -24,29 +24,28 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.kanban.codegen;
+package io.spine.validation;
 
-import com.google.common.collect.ImmutableSet;
-import io.spine.protodata.plugin.Plugin;
-import io.spine.protodata.plugin.Policy;
+import io.spine.protodata.TypeEntered;
+import io.spine.protodata.TypeName;
 import io.spine.protodata.plugin.ViewRepository;
+import io.spine.server.route.EventRouting;
 import org.jetbrains.annotations.NotNull;
 
+import static io.spine.server.route.EventRoute.withId;
+
 /**
- * A ProtoData plugin which attaches validation-related policies and views.
+ * A repository for the {@link MessageValidationView}.
+ *
+ * <p>Routes the {@code TypeEntered} events to the view by the type name.
  */
-@SuppressWarnings("unused") // Loaded by ProtoData via reflection.
-public class ValidationPlugin implements Plugin {
+class MessageValidationRepository
+        extends ViewRepository<TypeName, MessageValidationView, MessageValidation> {
 
-    @NotNull
     @Override
-    public ImmutableSet<Policy<?>> policies() {
-        return ImmutableSet.of(new RequiredRulePolicy());
-    }
-
-    @NotNull
-    @Override
-    public ImmutableSet<ViewRepository<?, ?, ?>> viewRepositories() {
-        return ImmutableSet.of(new MessageValidationRepository());
+    protected void setupEventRouting(@NotNull EventRouting<TypeName> routing) {
+        super.setupEventRouting(routing);
+        routing.route(TypeEntered.class,
+                      (message, context) -> withId(message.getType().getName()));
     }
 }
