+++ model/src/main/java/io/spine/validation/RequiredRulePolicy.java
@@ -24,12 +24,14 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.kanban.codegen;
+package io.spine.validation;
 
 import com.google.protobuf.DescriptorProtos.FieldOptions;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import com.google.protobuf.Extension;
 import io.spine.core.External;
+import io.spine.option.OptionsProto;
+import io.spine.protodata.Ast;
 import io.spine.protodata.Field;
 import io.spine.protodata.FieldOptionDiscovered;
 import io.spine.protodata.MessageType;
@@ -39,15 +41,8 @@ import io.spine.protodata.plugin.Policy;
 import io.spine.server.event.React;
 import io.spine.server.model.Nothing;
 import io.spine.server.tuple.EitherOf2;
-import io.spine.validation.RuleAdded;
-import io.spine.validation.SimpleRule;
-import io.spine.validation.Value;
+import io.spine.util.Exceptions;
 import org.jetbrains.annotations.NotNull;
-
-import static io.spine.option.OptionsProto.required;
-import static io.spine.protodata.Ast.typeUrl;
-import static io.spine.util.Exceptions.newIllegalArgumentException;
-import static io.spine.validation.ComparisonOperator.NOT_EQUAL;
 
 /**
  * A {@link Policy} which controls whether or not a field should be validated as {@code required}.
@@ -63,20 +58,20 @@ public final class RequiredRulePolicy extends Policy<FieldOptionDiscovered> {
     @React
     public EitherOf2<RuleAdded, Nothing> whenever(@External FieldOptionDiscovered event) {
         Option option = event.getOption();
-        if (isOption(option, required)) {
+        if (isOption(option, OptionsProto.required)) {
             ProtobufSourceFile file = select(ProtobufSourceFile.class)
                     .withId(event.getFile())
-                    .orElseThrow(() -> newIllegalArgumentException(
+                    .orElseThrow(() -> Exceptions.newIllegalArgumentException(
                             "Unknown file `%s`.", event.getFile()
                                                        .getValue()
                     ));
             MessageType type = file.getTypeMap()
-                                   .get(typeUrl(event.getType()));
+                                   .get(Ast.typeUrl(event.getType()));
             Field field = type.getFieldList()
                               .stream()
                               .filter(f -> f.getName().equals(event.getField()))
                               .findAny()
-                              .orElseThrow(() -> newIllegalArgumentException(
+                              .orElseThrow(() -> Exceptions.newIllegalArgumentException(
                                       "Unknown field `%s`.", event.getField()
                               ));
             return EitherOf2.withA(requiredRule(field));
@@ -90,7 +85,7 @@ public final class RequiredRulePolicy extends Policy<FieldOptionDiscovered> {
         SimpleRule rule = SimpleRule.newBuilder()
                                     .setErrorMessage("Field must be set.")
                                     .setField(field)
-                                    .setSign(NOT_EQUAL)
+                                    .setSign(ComparisonOperator.NOT_EQUAL)
                                     .setOtherValue(defaultValue)
                                     .vBuild();
         return RuleAdded.newBuilder()
