+++ pom.xml
@@ -12,7 +12,7 @@ all modules and does not describe the project structure per-subproject.
 
 <groupId>io.spine</groupId>
 <artifactId>validation</artifactId>
-<version>0.0.10</version>
+<version>0.0.11</version>
 
 <inceptionYear>2015</inceptionYear>
 
@@ -39,20 +39,26 @@ all modules and does not describe the project structure per-subproject.
   </dependency>
   <dependency>
     <groupId>io.spine.protodata</groupId>
+    <artifactId>codegen-java</artifactId>
+    <version>0.0.11</version>
+    <scope>compile</scope>
+  </dependency>
+  <dependency>
+    <groupId>io.spine.protodata</groupId>
     <artifactId>compiler</artifactId>
-    <version>0.0.10</version>
+    <version>0.0.11</version>
     <scope>compile</scope>
   </dependency>
   <dependency>
     <groupId>org.jboss.forge.roaster</groupId>
     <artifactId>roaster-api</artifactId>
-    <version>2.21.2.Final</version>
+    <version>2.22.2.Final</version>
     <scope>compile</scope>
   </dependency>
   <dependency>
     <groupId>org.jboss.forge.roaster</groupId>
     <artifactId>roaster-jdt</artifactId>
-    <version>2.21.2.Final</version>
+    <version>2.22.2.Final</version>
     <scope>compile</scope>
   </dependency>
   <dependency>
