+++ settings.gradle.kts
@@ -27,3 +27,4 @@
 rootProject.name = "validation"
 
 include("model")
+include("java")
